Add iotests
An io test is a simple .hissa and an accompanying .expected containing
the program's expected output.

At build time, the iotests tool will assemble and execute each .hissa
with an accompanying .expected file. It will fail if their outputs do
not match.

Also, remove CLI interface (for now) in favor of a library interface to
the VM. A new and improved CLI (based on this interface) is forthcoming.