Add workaround for compiler bug
The Zig compiler's "parameter reference optimization" can use
references when passing value types as arguments to avoid additional
copies.

In our case, this optimization is incorrectly applied to code that
copies a value from and then appends it to the VM stack. Growing an
ArrayList invalidates existing item pointers, so the reference is no
longer valid, and the code segfaults.

We can modify the code slightly so the PRO does not get applied.

https://github.com/ziglang/zig/issues/23050